pyproject.toml
[project]
name = "jemdzem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.16.1",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.4",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
